handle get_all_dependencies() exceptions
